Moved DependabotInstallationManager to it's own file.
Classes in PHP should be 1 per file. Dependabot is now using the
autoloader instead of manually requiring files in all cases.
