+++ helpers/php/composer.json
@@ -5,5 +5,10 @@
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^2.9"
+    },
+    "autoload": {
+        "psr-4": {
+            "\\Dependabot\\PHP\\": "src/"
+        }
     }
 }
